小修改

+++ goods/src/main/java/cn/edu/xmu/goods/controller/GoodsController.java
@@ -120,7 +120,7 @@ public class GoodsController {
             @ApiResponse(code = 0, message = "成功"),
     })
     @PostMapping("/shops/{shopId}/spus/{id}/skus")
-    public Object addSkutoSpu(@PathVariable Integer shopId,@PathVariable Integer id,@RequestBody SkuVo skuVo){
+    public Object addSkutoSpu(@PathVariable Long shopId,@PathVariable Long id,@RequestBody SkuVo skuVo){
         if(shopId==null||id==null){
             return Common.decorateReturnObject(new ReturnObject(ResponseCode.FIELD_NOTVALID));
         }
